Nicolas regnault: - change int to long in PrintStateMonomial - improve FQHESphereJackGenerator output options
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@2238 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/BosonOnSphereLong.cc
@@ -503,7 +503,7 @@ ostream& BosonOnSphereLong::PrintState (ostream& Str, int state)
 // state = ID of the state to print
 // return value = reference on current output stream 
 
-ostream& BosonOnSphereLong::PrintStateMonomial (ostream& Str, int state)
+ostream& BosonOnSphereLong::PrintStateMonomial (ostream& Str, long state)
 {
   unsigned long* TmpMonomial = new unsigned long [this->NbrBosons];
   this->ConvertToMonomial(this->FermionBasis->StateDescription[state], this->FermionBasis->StateLzMax[state], TmpMonomial);
