+++ FQHE/src/HilbertSpace/BosonOnTorusShort.h
@@ -190,7 +190,7 @@ class BosonOnTorusShort :  public ParticleOnTorus
   // Str = reference on current output stream 
   // state = ID of the state to print
   // return value = reference on current output stream 
-  virtual ostream& PrintStateMonomial (ostream& Str, int state);
+  virtual ostream& PrintStateMonomial (ostream& Str, long state);
   
   // apply Prod_i a_ni operator to a given state. Warning, the resulting state may not belong to the current Hilbert subspace. It will be keep in cache until next ProdA call
   //
