+++ FQHE/src/HilbertSpace/FermionOnSphereEfficient.cc
@@ -824,7 +824,7 @@ ostream& FermionOnSphereEfficient::PrintState (ostream& Str, int state)
 // state = ID of the state to print
 // return value = reference on current output stream 
 
-ostream& FermionOnSphereEfficient::PrintStateMonomial (ostream& Str, int state)
+ostream& FermionOnSphereEfficient::PrintStateMonomial (ostream& Str, long state)
 {
   unsigned long TmpState = this->StateDescription[state];
   Str << "[";
